// ppr-backend/src/domain/transactions/transaction-types.enum.ts
export enum TransactionTypes {
  ADD_CONTRIBUTION_PHASE = 'add_contribution_phase',
  CREATE_PROJECT = 'create_project',
  ADD_PHASE_PROJECT = 'add_phase_project',
  ADD_TASK_PHASE_PROJECT = 'add_task_phase_project',
  ADD_EVIDENCE_PROJECT = 'add_evidence_project',
  ADD_ORGANIZATION = 'add_organization',
  ADD_PHASE = 'add_phase',
  ADD_PROJECT = 'add_project',
  UPDATE_ORGANIZATION = 'update_organization',
  UPDATE_USER = 'update_user',
  UPDATE_PROJECT = 'update_project',
  UPDATE_CONTRIBUTION_RECEIVED_PROJECT = 'update_contribution_received_project',
  ADD_PHASE_PROJECT_TASK = 'add_phase_project_task',
  ADD_MEMBERS_PROJECT = 'add_members_project',
  ADD_USER_PROJECT = 'add_user_project',
  REGISTER_ACCESS_USER = 'register_access_user',
  ADD_TASK = 'add_task',
  ADD_EVIDENCE = 'add_evidence',
  UPDATE_PHASE_PROJECT = 'update_phase_project',
  UPDATE_PHASE_PROJECT_TASK = 'update_phase_project_task',
  IMPORT_EVIDENCES= 'import_evidences',
}